处理导出worktime到excel时的customer、outdoor等几个可能为null的字段导致主线程堵塞

+++ Wbs.Everdigm.Web/Wbs.Everdigm.Desktop/Handlers/PartialDataHandlerIridium.cs
@@ -395,7 +395,7 @@ namespace Wbs.Everdigm.Desktop
                 });
                 if (obj.TB_Terminal.Satellite != (int?)null)
                 {
-                    HandleIridiumMTFlow(obj.TB_Terminal.Satellite.Value, (int)(obj.Content.Length / 2));
+                    HandleIridiumMTFlow(obj.TB_Terminal.Satellite.Value, obj.Content.Length / 2);
                 }
             }
         }
